src/core/Application/Users/Queries/GetUserInfo/GetUserInfo.cs


using Application.Common.Exceptions;
using Application.Common.Interfaces;
using Application.Common.Models;
using Application.Common.Security;

namespace Application.Users.Queries.GetUserInfo
{
    [Authorize]
    public record GetUserInfoQuery : IRequest<UserDto>
    {
        public string UserName { get; set; } = null!;
    }

    internal class GetUserInfoQueryHandler : IRequestHandler<GetUserInfoQuery, UserDto>
    {
        private readonly IApplicationDbContext _context;
        private readonly IMapper _mapper;

        public GetUserInfoQueryHandler(IApplicationDbContext context, IMapper mapper)
        {
            _context = context;
            _mapper = mapper;
        }

        public async Task<UserDto> Handle(GetUserInfoQuery request, CancellationToken cancellationToken)
        {
            var user = await _context.Users.Where(x => x.UserName == request.UserName)
            .ProjectTo<UserDto>(_mapper.ConfigurationProvider)
            .FirstOrDefaultAsync();
            Guard.Against.NotFound(request.UserName, user);
            return user;
        }
    }
}